rgeowheels: Convenient Access to 'Python' 'Wheel' Packages Prepared for 'Windows'

Downloads pre-compiled 'Windows' 'wheel' files (.whl) for 'Python' geospatial packages from https://github.com/cgohlke/geospatial-wheels. These are unofficial binary installers for some geospatial libraries prepared by Christoph Gohlke. Wheels for various packages, 'Python' versions, and architectures are made available via 'GitHub' releases, providing the ability to revert to prior versions when needed.

See also